// src/ts/enums/HtmlSelectors.ts
export enum HtmlSelectors {
  WEIGHT = '#weight',
  HEIGHT = '#height',
  WAIST = '#waist',
  NECK = '#neck',
  HIP = '#hip',
  AGE = '#age',
  WEIGHT_GOAL = '#weightGoal',
  WEEKS_TO_WEIGHT_GOAL = '#weeksToWeightGoal',
  DAILY_CALORIES = '#dailyCalories',
  UNIT_SYSTEM = '#unitSystem',
  GENDER = '#gender',
  ACTIVITY_LEVEL = '#activityLevel',
  RESULT_TABLE = '.resultTable',
  ERROR_MESSAGE = '.errorMessage',
  FORM = 'form',
}